README
if you want to buy more than a thousand  of csgo cases? you just need to send to the market beforce your stocks be fulled. and it can also remove the case auto if you wang to open the case or sell the case in suitable price.
First,  open the cmd in the local of the chrome and in put 'chrome.exe --remote-debugging-port=9527 --user-data-dir=“C:\selenium1\AutomationProfile'  to open the chrome with debug mode.
Then sign in the steam account and then run the code.
it will open a excel to statistics the money and the case you send the price and nums.